Updated indexing. Changed FabType property to byte.

+++ Solution/Fabric.Api.Traversal/Steps/Nodes/RootStep.cs
@@ -1,6 +1,6 @@
 ﻿// GENERATED CODE
 // Changes made to this source file will be overwritten
-// Generated on 5/5/2013 9:59:20 PM
+// Generated on 5/16/2013 10:09:03 PM
 
 using System.Collections.Generic;
 using System.Linq;
@@ -68,7 +68,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IAppStep ContainsAppList {
 			get {
 				var step = new AppStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.App));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.App));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -78,7 +78,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IClassStep ContainsClassList {
 			get {
 				var step = new ClassStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.Class));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.Class));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -88,7 +88,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IInstanceStep ContainsInstanceList {
 			get {
 				var step = new InstanceStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.Instance));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.Instance));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -98,7 +98,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IMemberStep ContainsMemberList {
 			get {
 				var step = new MemberStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.Member));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.Member));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -108,7 +108,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IMemberTypeAssignStep ContainsMemberTypeAssignList {
 			get {
 				var step = new MemberTypeAssignStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.MemberTypeAssign));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.MemberTypeAssign));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -118,7 +118,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IUrlStep ContainsUrlList {
 			get {
 				var step = new UrlStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.Url));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.Url));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -128,7 +128,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IUserStep ContainsUserList {
 			get {
 				var step = new UserStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.User));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.User));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
@@ -138,7 +138,7 @@ namespace Fabric.Api.Traversal.Steps.Nodes {
 		public IFactorStep ContainsFactorList {
 			get {
 				var step = new FactorStep(Path);
-				string p = Path.AddParam(new WeaverQueryVal((int)NodeFabType.Factor));
+				string p = Path.AddParam(new WeaverQueryVal((byte)NodeFabType.Factor));
 				Path.AddSegment(step, "V('"+PropDbName.Node_FabType+"',"+p+")");
 				return step;
 			}
